+++ frontend/src/pages/inq/Inq.js
@@ -1,3 +1,5 @@
+import React from 'react';
+import Header from '../../components/mocules/Header';
 import { AutoFocusPlugin } from '@lexical/react/LexicalAutoFocusPlugin';
 import { LexicalComposer } from '@lexical/react/LexicalComposer';
 import { ContentEditable } from '@lexical/react/LexicalContentEditable';
@@ -5,7 +7,7 @@ import { LexicalErrorBoundary } from '@lexical/react/LexicalErrorBoundary';
 import { HistoryPlugin } from '@lexical/react/LexicalHistoryPlugin';
 import { RichTextPlugin } from '@lexical/react/LexicalRichTextPlugin';
 
-import ToolbarPlugin from '../../plugins/ToolbarPlugin'
+import ToolbarPlugin from '../../plugins/ToolbarPlugin';
 import TreeViewPlugin from '../../plugins/TreeViewPlugin';
 import Theme from './Theme';
 
@@ -24,18 +26,23 @@ const editorConfig = {
     theme: Theme,
 };
 
-export default function Inq() {
+function Inq() {
     return (
-        <LexicalComposer initialConfig={editorConfig}>
-            <div className="editor-container">
-                <ToolbarPlugin />
-                <div className="editor-inner">
-                    <RichTextPlugin contentEditable={<ContentEditable className="editor-input" aria-placeholder={placeholder} placeholder={<div className="editor-placeholder">{placeholder}</div>} />} ErrorBoundary={LexicalErrorBoundary} />
-                    <HistoryPlugin />
-                    <AutoFocusPlugin />
-                    <TreeViewPlugin />
+        <>
+            <Header login={true} />
+            <LexicalComposer initialConfig={editorConfig}>
+                <div className="editor-container">
+                    <ToolbarPlugin />
+                    <div className="editor-inner">
+                        <RichTextPlugin contentEditable={<ContentEditable className="editor-input" aria-placeholder={placeholder} placeholder={<div className="editor-placeholder">{placeholder}</div>} />} ErrorBoundary={LexicalErrorBoundary} />
+                        <HistoryPlugin />
+                        <AutoFocusPlugin />
+                        <TreeViewPlugin />
+                    </div>
                 </div>
-            </div>
-        </LexicalComposer>
+            </LexicalComposer>
+        </>
     );
 }
+
+export default Inq;
